パスワード変更実装

+++ routes/web.php
@@ -37,7 +37,7 @@ Route::get('/', function () {
 
 Route::get('/welcome', function () {
     return view('welcome');
-})->name('welcome');
+})->name('welcome')->middleware(['auth']);
 
 // ここのルートを変更して各自の画面を確認する
 
@@ -84,7 +84,7 @@ Route::post('/logout', [AuthController::class, 'logout'])->name('logout');
 Route::get('/password/change', [PasswordController::class, 'showChangeForm'])->name('password.change.form');
 Route::post('/password/change', [PasswordController::class, 'changePassword'])->name('password.change');
 
-// パスワードリセットのためのルート定義
+// パスワードリセットのためのルート定義(使用できない)
 Route::get('/password/reset', [ForgotPasswordController::class, 'showLinkRequestForm'])->name('password.request');
 Route::post('/password/email', [ForgotPasswordController::class, 'sendResetLinkEmail'])->name('password.email');
 Route::get('/password/reset/{token}', [ResetPasswordController::class, 'showResetForm'])->name('password.reset');
@@ -92,9 +92,9 @@ Route::post('/password/reset', [ResetPasswordController::class, 'reset'])->name(
 
 // admin.dashboardのルート
 Route::get('/admin', [UserController::class, 'index'])->name('admin.dashboard')->middleware(['auth', 'can:admin']);
-Route::get('/admin/{user}/edit', [UserController::class, 'edit'])->name('user.edit');
-Route::post('/admin/{user}', [UserController::class, 'update'])->name('user.update');
-Route::delete('/admin/{user}', [UserController::class, 'destroy'])->name('user.destroy');
+Route::get('/admin/{user}/edit', [UserController::class, 'edit'])->name('user.edit')->middleware(['auth', 'can:admin']);
+Route::post('/admin/{user}', [UserController::class, 'update'])->name('user.update')->middleware(['auth', 'can:admin']);
+Route::delete('/admin/{user}', [UserController::class, 'destroy'])->name('user.destroy')->middleware(['auth', 'can:admin']);
 
 // --------------------------------------------------------------
